+++ lib/src/bones_api_entity.dart
@@ -145,6 +145,10 @@ typedef EntityCache = JsonEntityCache;
 /// Rules to resolve entities.
 /// Used by [EntityHandler] and [DBEntityRepository].
 class EntityResolutionRules {
+  /// A `const` instance without any resolution rules to apply.
+  static const EntityResolutionRules innocuous =
+      _EntityResolutionRulesInnocuous();
+
   final bool? _allowEntityFetch;
 
   /// If `true` it will allow the use of on DB/repository to fetch an entity by an ID reference.
@@ -216,6 +220,43 @@ class EntityResolutionRules {
         allLazy = true,
         allEager = false;
 
+  /// Returns `true` if this instance is equivalent to [innocuous] instance (no resolution rules to apply).
+  bool get isInnocuous =>
+      _allowEntityFetch == null &&
+      !allowReadFile &&
+      (eagerEntityTypes == null || eagerEntityTypes!.isEmpty) &&
+      (lazyEntityTypes == null || lazyEntityTypes!.isEmpty) &&
+      allLazy == null &&
+      allEager == null;
+
+  bool get isValid => _validateImpl() == null;
+
+  void validate() {
+    var error = _validateImpl();
+    if (error != null) {
+      throw ValidateEntityResolutionRulesError(this, error);
+    }
+  }
+
+  String? _validateImpl() {
+    var eagerEntityTypes = this.eagerEntityTypes;
+    var lazyEntityTypes = this.lazyEntityTypes;
+
+    if (eagerEntityTypes != null &&
+        lazyEntityTypes != null &&
+        eagerEntityTypes.isNotEmpty &&
+        lazyEntityTypes.isNotEmpty) {
+      if (eagerEntityTypes.any((t) => lazyEntityTypes.contains(t))) {
+        var conflict =
+            eagerEntityTypes.where((t) => lazyEntityTypes.contains(t)).toList();
+
+        return "Conflicting `eagerEntityTypes` and `lazyEntityTypes`: $conflict";
+      }
+    }
+
+    return null;
+  }
+
   /// Returns `true` if [entityType] load is eager.
   /// - [def] is returned in case there's not rule for [entityType].
   bool isEagerEntityType(Type entityType, [bool def = false]) {
@@ -270,14 +311,242 @@ class EntityResolutionRules {
     return entityType != null ? isLazyEntityType(entityType, def) : false;
   }
 
+  /// Copies this instance, allowing fields overwrite.
+  EntityResolutionRules copyWith(
+      {bool? allowEntityFetch,
+      bool? allowReadFile,
+      List<Type>? lazyEntityTypes,
+      List<Type>? eagerEntityTypes,
+      bool? allLazy,
+      bool? allEager}) {
+    var resolutionRules = EntityResolutionRules(
+      allowEntityFetch: allowEntityFetch ?? _allowEntityFetch,
+      allowReadFile: allowReadFile ?? this.allowReadFile,
+      lazyEntityTypes: lazyEntityTypes ?? this.lazyEntityTypes,
+      eagerEntityTypes: eagerEntityTypes ?? this.eagerEntityTypes,
+      allLazy: allLazy ?? this.allLazy,
+      allEager: allEager ?? this.allEager,
+    );
+    return resolutionRules.isInnocuous ? innocuous : resolutionRules;
+  }
+
+  /// Merges this instances with [other].
+  EntityResolutionRules merge(EntityResolutionRules? other) {
+    if (other == null || other.isInnocuous) {
+      return isInnocuous ? innocuous : this;
+    } else if (isInnocuous) {
+      return other;
+    }
+
+    var allowEntityFetch = _allowEntityFetch;
+    if (other._allowEntityFetch != null) {
+      if (allowEntityFetch == null) {
+        allowEntityFetch = other._allowEntityFetch;
+      } else if (allowEntityFetch != other._allowEntityFetch) {
+        throw MergeEntityResolutionRulesError(this, other, 'allowEntityFetch');
+      }
+    }
+
+    var allowReadFile = this.allowReadFile || other.allowReadFile;
+
+    var allLazy = this.allLazy;
+    if (other.allLazy != null) {
+      if (allLazy == null) {
+        allLazy = other.allLazy;
+      } else if (allLazy != other.allLazy) {
+        throw MergeEntityResolutionRulesError(this, other, 'allLazy');
+      }
+    }
+
+    var allEager = this.allEager;
+    if (other.allEager != null) {
+      if (allEager == null) {
+        allEager = other.allEager;
+      } else if (allEager != other.allEager) {
+        throw MergeEntityResolutionRulesError(this, other, 'allEager');
+      }
+    }
+
+    var lazyEntityTypes = this.lazyEntityTypes;
+
+    var otherLazyEntityTypes = other.lazyEntityTypes;
+    if (otherLazyEntityTypes != null && otherLazyEntityTypes.isNotEmpty) {
+      lazyEntityTypes = [...?lazyEntityTypes, ...otherLazyEntityTypes];
+    }
+
+    var eagerEntityTypes = this.eagerEntityTypes;
+
+    var otherEagerEntityTypes = other.eagerEntityTypes;
+    if (otherEagerEntityTypes != null && otherEagerEntityTypes.isNotEmpty) {
+      eagerEntityTypes = [...?eagerEntityTypes, ...otherEagerEntityTypes];
+    }
+
+    var merge = EntityResolutionRules(
+      allowEntityFetch: allowEntityFetch,
+      allowReadFile: allowReadFile,
+      lazyEntityTypes: lazyEntityTypes,
+      eagerEntityTypes: eagerEntityTypes,
+      allLazy: allLazy,
+      allEager: allEager,
+    );
+
+    merge.validate();
+
+    return merge;
+  }
+
   @override
   String toString() {
-    return 'EntityResolutionRules{allowEntityFetch: $allowEntityFetch, allowReadFile: $allowReadFile, lazyEntityTypes: $lazyEntityTypes, eagerEntityTypes: $eagerEntityTypes}';
+    final lazyEntityTypes = this.lazyEntityTypes;
+    final eagerEntityTypes = this.eagerEntityTypes;
+
+    return isInnocuous
+        ? 'EntityResolutionRules{innocuous}'
+        : 'EntityResolutionRules{${[
+            if (allLazy != null) 'allLazy: $allLazy',
+            if (allEager != null) 'allEager: $allEager',
+            if (allowEntityFetch) 'allowEntityFetch',
+            if (allowReadFile) 'allowReadFile',
+            if (lazyEntityTypes != null && lazyEntityTypes.isNotEmpty)
+              'lazyEntityTypes: $lazyEntityTypes',
+            if (eagerEntityTypes != null && eagerEntityTypes.isNotEmpty)
+              'eagerEntityTypes: $eagerEntityTypes',
+          ].join(', ')}'
+            '}';
+  }
+}
+
+class ValidateEntityResolutionRulesError extends Error {
+  final EntityResolutionRules resolutionRules;
+  final String message;
+
+  ValidateEntityResolutionRulesError(this.resolutionRules, this.message);
+
+  @override
+  String toString() =>
+      "Error validating `EntityResolutionRules`: $message >> $resolutionRules";
+}
+
+/// [EntityResolutionRules.merge] error.
+class MergeEntityResolutionRulesError extends Error {
+  final EntityResolutionRules a;
+  final EntityResolutionRules b;
+
+  final String conflict;
+
+  MergeEntityResolutionRulesError(this.a, this.b, this.conflict);
+
+  @override
+  String toString() =>
+      "Can't merge `EntityResolutionRules`! Conflict: $conflict >> $a <!> $b";
+}
+
+class _EntityResolutionRulesInnocuous implements EntityResolutionRules {
+  const _EntityResolutionRulesInnocuous();
+
+  @override
+  bool get isInnocuous => true;
+
+  @override
+  bool get isValid => true;
+
+  @override
+  void validate() {}
+
+  @override
+  String? _validateImpl() => null;
+
+  @override
+  bool? get _allowEntityFetch => null;
+
+  @override
+  bool? get allEager => null;
+
+  @override
+  bool? get allLazy => null;
+
+  @override
+  bool get allowEntityFetch => false;
+
+  @override
+  bool get allowReadFile => false;
+
+  @override
+  List<Type>? get eagerEntityTypes => null;
+
+  @override
+  List<Type>? get lazyEntityTypes => null;
+
+  @override
+  bool isEagerEntityType(Type entityType, [bool def = false]) => false;
+
+  @override
+  bool isEagerEntityTypeInfo(TypeInfo entityTypeInfo, [bool def = false]) =>
+      false;
+
+  @override
+  bool isLazyEntityType(Type entityType, [bool def = false]) => false;
+
+  @override
+  bool isLazyEntityTypeInfo(TypeInfo entityTypeInfo, [bool def = false]) =>
+      false;
+
+  @override
+  EntityResolutionRules copyWith(
+          {bool? allowEntityFetch,
+          bool? allowReadFile,
+          List<Type>? lazyEntityTypes,
+          List<Type>? eagerEntityTypes,
+          bool? allLazy,
+          bool? allEager}) =>
+      EntityResolutionRules(
+        allowEntityFetch: allowEntityFetch,
+        allowReadFile: allowReadFile ?? false,
+        lazyEntityTypes: lazyEntityTypes,
+        eagerEntityTypes: eagerEntityTypes,
+        allLazy: allLazy,
+        allEager: allEager,
+      );
+
+  @override
+  EntityResolutionRules merge(EntityResolutionRules? other) {
+    if (other != null) {
+      return other.isInnocuous ? EntityResolutionRules.innocuous : other;
+    } else {
+      return this;
+    }
+  }
+
+  @override
+  String toString() => 'EntityResolutionRules{innocuous}';
+}
+
+/// Mixin to resolve the [EntityResolutionRules] to apply.
+mixin EntityRulesResolver {
+  /// The current [EntityResolutionRules] of the current context.
+  EntityResolutionRules? getContextEntityResolutionRules() {
+    return null;
+  }
+
+  /// Resolves the [resolutionRules] to apply. Merges with the current
+  /// [EntityResolutionRules] context if needed.
+  ///
+  /// See [getContextEntityResolutionRules] and [EntityResolutionRules.merge].
+  EntityResolutionRules resolveEntityResolutionRules(
+      EntityResolutionRules? resolutionRules) {
+    var context = getContextEntityResolutionRules();
+
+    if (resolutionRules == null || resolutionRules.isInnocuous) {
+      return context ?? EntityResolutionRules.innocuous;
+    } else {
+      var merge = resolutionRules.merge(context);
+      return merge;
+    }
   }
 }
 
 /// Base class to implement entities handlers.
-abstract class EntityHandler<O> with FieldsFromMap {
+abstract class EntityHandler<O> with FieldsFromMap, EntityRulesResolver {
   /// The provider of [EntityHandler] instances for different [Type]s.
   final EntityHandlerProvider provider;
 
@@ -563,6 +832,9 @@ abstract class EntityHandler<O> with FieldsFromMap {
       EntityResolutionRules? resolutionRules}) {
     entityCache ??= JsonEntityCacheSimple();
 
+    final resolutionRulesF =
+        resolutionRules = resolveEntityResolutionRules(resolutionRules);
+
     var fieldsTypes = this.fieldsTypes(o);
 
     var resolved = fields.map((f, v) {
@@ -572,7 +844,7 @@ abstract class EntityHandler<O> with FieldsFromMap {
           entityCache: entityCache,
           entityHandlerProvider: entityHandlerProvider,
           entityRepositoryProvider: entityRepositoryProvider,
-          resolutionRules: resolutionRules);
+          resolutionRules: resolutionRulesF);
 
       if (t != null) {
         if (t.isEntityReferenceType) {
@@ -663,6 +935,9 @@ abstract class EntityHandler<O> with FieldsFromMap {
       return value as T?;
     }
 
+    final resolutionRulesF =
+        resolutionRules = resolveEntityResolutionRules(resolutionRules);
+
     var tType = TypeInfo.from(T);
     var valueType = value != null ? TypeInfo.from(value) : null;
 
@@ -679,7 +954,7 @@ abstract class EntityHandler<O> with FieldsFromMap {
     } else if (type.equalsType(_typeInfoDynamicInt)) {
       return DynamicInt.from(value) as T?;
     } else if (type.equalsType(_typeInfoUint8List)) {
-      return _resolveValueAsUInt8List(value, resolutionRules) as T?;
+      return _resolveValueAsUInt8List(value, resolutionRulesF) as T?;
     }
 
     entityCache ??= JsonEntityCacheSimple();
@@ -706,7 +981,7 @@ abstract class EntityHandler<O> with FieldsFromMap {
               ? valEntityHandler.createFromMap(value,
                   entityProvider: entityProvider,
                   entityCache: entityCache,
-                  resolutionRules: resolutionRules)
+                  resolutionRules: resolutionRulesF)
               : value;
         }
 
@@ -719,7 +994,7 @@ abstract class EntityHandler<O> with FieldsFromMap {
             elementType, entityHandlerProvider, entityRepositoryProvider);
 
         return _resolveListValueByType<T>(valEntityHandler, value, elementType,
-            entityProvider, entityCache, resolutionRules);
+            entityProvider, entityCache, resolutionRulesF);
       } else if (type.isEntityReferenceListType) {
         var elementType = type.arguments.first;
 
@@ -727,7 +1002,7 @@ abstract class EntityHandler<O> with FieldsFromMap {
             elementType, entityHandlerProvider, entityRepositoryProvider);
 
         var eagerEntityType =
-            resolutionRules?.isEagerEntityType(elementType.type) ?? false;
+            resolutionRulesF.isEagerEntityType(elementType.type);
         if (!eagerEntityType) {
           return elementType.toEntityReferenceList(value,
               entityProvider: entityProvider,
@@ -735,7 +1010,7 @@ abstract class EntityHandler<O> with FieldsFromMap {
         }
 
         return _resolveListValueByType<T>(valEntityHandler, value, elementType,
-            entityProvider, entityCache, resolutionRules);
+            entityProvider, entityCache, resolutionRulesF);
       } else {
         return type.parseEntity<T>(value);
       }
@@ -751,12 +1026,11 @@ abstract class EntityHandler<O> with FieldsFromMap {
       if (parsed != null) return parsed;
 
       if (entityProvider != null && entityType != null) {
-        var eagerEntityType =
-            resolutionRules?.isEagerEntityType(entityType) ?? false;
+        var eagerEntityType = resolutionRulesF.isEagerEntityType(entityType);
 
         if (eagerEntityType) {
           var entityAsync = entityProvider.getEntityByID(value,
-              type: entityType, resolutionRules: resolutionRules);
+              type: entityType, resolutionRules: resolutionRulesF);
 
           if (entityAsync != null) {
             return entityAsync.resolveMapped<T?>((entity) {
@@ -771,7 +1045,7 @@ abstract class EntityHandler<O> with FieldsFromMap {
                   entityCache!,
                   entityHandlerProvider,
                   entityRepositoryProvider,
-                  resolutionRules);
+                  resolutionRulesF);
             });
           }
         }
@@ -784,7 +1058,7 @@ abstract class EntityHandler<O> with FieldsFromMap {
           entityCache,
           entityHandlerProvider,
           entityRepositoryProvider,
-          resolutionRules);
+          resolutionRulesF);
     } else {
       return type.parseEntity<T>(value);
     }
@@ -796,7 +1070,7 @@ abstract class EntityHandler<O> with FieldsFromMap {
       TypeInfo elementType,
       EntityProvider? entityProvider,
       EntityCache? entityCache,
-      EntityResolutionRules? resolutionRules) {
+      EntityResolutionRules resolutionRules) {
     if (!elementType.isBasicType) {
       var totalEntitiesToResolve = 0;
       var totalResolvedEntities = 0;
@@ -854,7 +1128,7 @@ abstract class EntityHandler<O> with FieldsFromMap {
   }
 
   FutureOr<Uint8List?> _resolveValueAsUInt8List(
-      Object? value, EntityResolutionRules? resolutionRules) {
+      Object? value, EntityResolutionRules resolutionRules) {
     if (value == null) return null;
     var bytes = TypeParser.parseUInt8List(value);
     if (bytes != null) return bytes;
@@ -868,7 +1142,7 @@ abstract class EntityHandler<O> with FieldsFromMap {
         var dataURL = DataURLBase64.parse(value);
         return dataURL?.payloadArrayBuffer;
       } else if (value.startsWith("url(") && value.endsWith(")")) {
-        var allowReadFile = resolutionRules?.allowReadFile ?? false;
+        var allowReadFile = resolutionRules.allowReadFile;
         if (!allowReadFile) return null;
 
         var path = value.substring(4, value.length - 1);
@@ -912,7 +1186,7 @@ abstract class EntityHandler<O> with FieldsFromMap {
       EntityCache entityCache,
       EntityHandlerProvider? entityHandlerProvider,
       EntityRepositoryProvider? entityRepositoryProvider,
-      EntityResolutionRules? resolutionRules) {
+      EntityResolutionRules resolutionRules) {
     var valEntityHandler = _resolveEntityHandler(
         type, entityHandlerProvider, entityRepositoryProvider);
 
@@ -924,14 +1198,13 @@ abstract class EntityHandler<O> with FieldsFromMap {
       }
     }
 
-    var allowEntityFetch = (resolutionRules?.allowEntityFetch ?? false) ||
-        entityCache.allowEntityFetch;
+    var allowEntityFetch =
+        resolutionRules.allowEntityFetch || entityCache.allowEntityFetch;
 
     if (allowEntityFetch && type.isEntityReferenceBaseType) {
       var entityType = type.entityType;
-      var eager = entityType != null &&
-          resolutionRules != null &&
-          resolutionRules.isEagerEntityType(entityType);
+      var eager =
+          entityType != null && resolutionRules.isEagerEntityType(entityType);
       allowEntityFetch = eager;
     }
 
@@ -3099,7 +3372,7 @@ extension EntityRepositoryProviderExtension on EntityRepositoryProvider {
 }
 
 abstract class EntityRepository<O extends Object> extends EntityAccessor<O>
-    with Initializable, Closable
+    with Initializable, Closable, EntityRulesResolver
     implements EntitySource<O>, EntityStorage<O> {
   static FutureOr<Map<String, dynamic>> resolveSubEntitiesFields(
       Map<String, dynamic> fields, Map<String, Type> subEntitiesFields,
@@ -4683,7 +4956,9 @@ extension TransactionOperationTypeExtension on TransactionOperationType {
   }
 }
 
-class TransactionEntityProvider implements EntityProvider {
+class TransactionEntityProvider
+    with EntityRulesResolver
+    implements EntityProvider {
   final Transaction transaction;
 
   final EntityRepositoryProvider entityRepositoryProvider;
@@ -4705,9 +4980,10 @@ class TransactionEntityProvider implements EntityProvider {
         return null;
       }
 
-      resolutionRules ??= this.resolutionRules;
+      final resolutionRulesF = resolutionRules =
+          resolveEntityResolutionRules(resolutionRules ?? this.resolutionRules);
 
-      var allowEntityFetch = resolutionRules?.allowEntityFetch ?? false;
+      var allowEntityFetch = resolutionRulesF.allowEntityFetch;
 
       if (!allowEntityFetch) {
         return null;
@@ -4718,7 +4994,7 @@ class TransactionEntityProvider implements EntityProvider {
       if (entityRepository == null) return null;
 
       var sel =
-          entityRepository.selectByID(id, resolutionRules: resolutionRules);
+          entityRepository.selectByID(id, resolutionRules: resolutionRulesF);
       return sel.resolveMapped((o) => o as O?);
     });
   }